Make types slightly more precise.

+++ pyhedonic/HedonicGameImpl.py
@@ -33,10 +33,10 @@ from numba import config, njit
 config.DISABLE_JIT = False
 
 # Unfortunately, specifying the shape of the array in the type hint does not work well
-type IntArray2D = npt.NDArray[np.integer]
+type IntArray2D = npt.NDArray[np.int_]
 
 # Unfortunately, specifying the shape of the array in the type hint does not work well
-type IntArray1D = npt.NDArray[np.integer]
+type IntArray1D = npt.NDArray[np.int_]
 
 type Game = IntArray2D
